bunch of changes
fixed all bugs so far and finished implementing the signing algorithm. Needs some refactoring though. Sha256 was having some issues with pre-processed data. I gave up on the point compression and decompression because I need to use some super complex algorithm for modular square root to decompress a point so public keys are just gonna be the full point (might compress it to remove the prime number base which is always the same)